backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
click==8.1.3
colorama==0.4.6
Django==4.2
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
joblib==1.2.0
nltk==3.8.1
pytz==2023.3
regex==2023.3.23
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
